fix: make sealed ledger mode explicit
