--- .github/scripts/apply_pr198.py
@@ -1,260 +1,0 @@
-from __future__ import annotations
-
-import json
-from pathlib import Path
-
-
-def replace_once(source: str, old: str, new: str, *, seam: str) -> str:
-    count = source.count(old)
-    if count != 1:
-        raise SystemExit(f"{seam} changed: expected one exact match, got {count}")
-    return source.replace(old, new)
-
-
-config_path = Path("trade_rl/workflows/market_walk_forward_config.py")
-source = config_path.read_text(encoding="utf-8")
-source = replace_once(
-    source,
-    "from dataclasses import asdict, dataclass\nfrom pathlib import Path\n",
-    "from dataclasses import asdict, dataclass\nfrom enum import StrEnum\nfrom pathlib import Path\n",
-    seam="config imports",
-)
-source = replace_once(
-    source,
-    "from trade_rl.workflows.walk_forward import WalkForwardWorkflowConfig\n\n\ndef _mapping",
-    '''from trade_rl.workflows.walk_forward import WalkForwardWorkflowConfig
-
-
-class SealedTestLedgerMode(StrEnum):
-    """Explicit trust semantics for outer-test access authorization."""
-
-    LOCAL_EXPLORATORY = "local_exploratory"
-    DURABLE_POSTGRES = "durable_postgres"
-
-
-def _mapping''',
-    seam="ledger mode declaration",
-)
-source = replace_once(
-    source,
-    '''    checkpoint_finalists_per_seed: int = 1
-    execution_sensitivity: ExecutionSensitivityConfig = ExecutionSensitivityConfig()
-    signal_digest: str = ""
-''',
-    '''    checkpoint_finalists_per_seed: int = 1
-    execution_sensitivity: ExecutionSensitivityConfig = ExecutionSensitivityConfig()
-    sealed_test_ledger_mode: SealedTestLedgerMode = (
-        SealedTestLedgerMode.LOCAL_EXPLORATORY
-    )
-    signal_digest: str = ""
-''',
-    seam="config field",
-)
-source = replace_once(
-    source,
-    '''    def __post_init__(self) -> None:
-        if not self.candidates:
-''',
-    '''    def __post_init__(self) -> None:
-        if not isinstance(self.sealed_test_ledger_mode, SealedTestLedgerMode):
-            raise ValueError("sealed_test_ledger_mode must be a supported mode")
-        if not self.candidates:
-''',
-    seam="config validation",
-)
-source = replace_once(
-    source,
-    '''        signal = payload.get("signal_digest", "")
-        if not isinstance(signal, str):
-            raise ValueError("signal_digest must be a string")
-        finalists_per_seed = payload.get("checkpoint_finalists_per_seed", 1)
-''',
-    '''        signal = payload.get("signal_digest", "")
-        if not isinstance(signal, str):
-            raise ValueError("signal_digest must be a string")
-        raw_ledger_mode = payload.get(
-            "sealed_test_ledger_mode",
-            SealedTestLedgerMode.LOCAL_EXPLORATORY.value,
-        )
-        if not isinstance(raw_ledger_mode, str):
-            raise ValueError("sealed_test_ledger_mode must be a string")
-        try:
-            sealed_test_ledger_mode = SealedTestLedgerMode(raw_ledger_mode)
-        except ValueError as error:
-            raise ValueError("sealed_test_ledger_mode is unsupported") from error
-        finalists_per_seed = payload.get("checkpoint_finalists_per_seed", 1)
-''',
-    seam="config parser",
-)
-source = replace_once(
-    source,
-    '''            checkpoint_finalists_per_seed=finalists_per_seed,
-            execution_sensitivity=sensitivity,
-            signal_digest=signal,
-''',
-    '''            checkpoint_finalists_per_seed=finalists_per_seed,
-            execution_sensitivity=sensitivity,
-            sealed_test_ledger_mode=sealed_test_ledger_mode,
-            signal_digest=signal,
-''',
-    seam="config construction",
-)
-source = replace_once(
-    source,
-    '''            "maximum_selection_drawdown": self.maximum_selection_drawdown,
-            "schema_version": self.schema_version,
-''',
-    '''            "maximum_selection_drawdown": self.maximum_selection_drawdown,
-            "sealed_test_ledger_mode": self.sealed_test_ledger_mode.value,
-            "schema_version": self.schema_version,
-''',
-    seam="config identity",
-)
-source = replace_once(
-    source,
-    '''    "MarketWalkForwardConfig",
-    "NamedCandidateRun",
-]
-''',
-    '''    "MarketWalkForwardConfig",
-    "NamedCandidateRun",
-    "SealedTestLedgerMode",
-]
-''',
-    seam="config exports",
-)
-config_path.write_text(source, encoding="utf-8")
-
-workflow_path = Path("trade_rl/workflows/market_walk_forward.py")
-source = workflow_path.read_text(encoding="utf-8")
-source = replace_once(
-    source,
-    '''from trade_rl.evaluation.walk_forward.folds import IndexRange, WalkForwardFold
-from trade_rl.integrations.checkpoints import StableBaselines3CheckpointLoader
-''',
-    '''from trade_rl.evaluation.walk_forward.folds import IndexRange, WalkForwardFold
-from trade_rl.evaluation.walk_forward.sealed_test import (
-    SealedTestLedger,
-    SealedTestLedgerProtocol,
-)
-from trade_rl.integrations.checkpoints import StableBaselines3CheckpointLoader
-''',
-    seam="workflow ledger imports",
-)
-source = replace_once(
-    source,
-    '''from trade_rl.workflows.market_walk_forward_config import (
-    MarketWalkForwardConfig as MarketWalkForwardConfig,
-)
-''',
-    '''from trade_rl.workflows.market_walk_forward_config import (
-    MarketWalkForwardConfig as MarketWalkForwardConfig,
-    SealedTestLedgerMode,
-)
-''',
-    seam="workflow mode import",
-)
-source = replace_once(
-    source,
-    '''def _sealed_test_ledger() -> PostgresSealedTestLedger | None:
-    database_url = os.environ.get("TRADE_RL_DATABASE_URL")
-    if not database_url:
-        return None
-    catalog = PostgresArtifactCatalog(database_url)
-    catalog.migrate()
-    return PostgresSealedTestLedger(catalog)
-''',
-    '''def _sealed_test_ledger(mode: SealedTestLedgerMode) -> SealedTestLedgerProtocol:
-    if mode is SealedTestLedgerMode.LOCAL_EXPLORATORY:
-        return SealedTestLedger()
-    database_url = os.environ.get("TRADE_RL_DATABASE_URL", "").strip()
-    if not database_url:
-        raise ValueError(
-            "durable PostgreSQL sealed-test ledger requires "
-            "TRADE_RL_DATABASE_URL"
-        )
-    catalog = PostgresArtifactCatalog(database_url)
-    return PostgresSealedTestLedger(catalog)
-''',
-    seam="ledger factory",
-)
-source = replace_once(
-    source,
-    "            sealed_test_ledger=_sealed_test_ledger(),\n",
-    "            sealed_test_ledger=_sealed_test_ledger(config.sealed_test_ledger_mode),\n",
-    seam="ledger injection",
-)
-source = replace_once(
-    source,
-    '''            "dataset_id": dataset.dataset_id,
-            "evaluation_digest": result.evaluation_digest,
-''',
-    '''            "dataset_id": dataset.dataset_id,
-            "evidence_tier": (
-                "durable_sealed"
-                if config.sealed_test_ledger_mode
-                is SealedTestLedgerMode.DURABLE_POSTGRES
-                else "exploratory_process_local"
-            ),
-            "evaluation_digest": result.evaluation_digest,
-''',
-    seam="evidence tier",
-)
-source = replace_once(
-    source,
-    '''            "production_status": "NO-GO",
-            "schema_version": "market_walk_forward_run_v5_deployable_ensemble",
-''',
-    '''            "production_status": "NO-GO",
-            "sealed_test_ledger_durable": (
-                config.sealed_test_ledger_mode
-                is SealedTestLedgerMode.DURABLE_POSTGRES
-            ),
-            "sealed_test_ledger_mode": config.sealed_test_ledger_mode.value,
-            "schema_version": "market_walk_forward_run_v5_deployable_ensemble",
-''',
-    seam="ledger evidence",
-)
-workflow_path.write_text(source, encoding="utf-8")
-
-example_modes = {
-    Path("examples/binance/walk-forward-smoke.json"): "local_exploratory",
-    Path("examples/binance-multitimeframe/walk-forward-full.json"): "durable_postgres",
-    Path("examples/binance-multitimeframe/walk-forward-growth-optimal.json"): "durable_postgres",
-}
-for path, mode in example_modes.items():
-    payload = json.loads(path.read_text(encoding="utf-8"))
-    payload["sealed_test_ledger_mode"] = mode
-    path.write_text(
-        json.dumps(payload, ensure_ascii=False, indent=2) + "\n",
-        encoding="utf-8",
-    )
-
-test_path = Path("tests/workflows/test_explicit_sealed_ledger_mode.py")
-source = test_path.read_text(encoding="utf-8")
-addition = '''
-
-@pytest.mark.parametrize(
-    ("path", "expected"),
-    (
-        ("examples/binance/walk-forward-smoke.json", "local_exploratory"),
-        (
-            "examples/binance-multitimeframe/walk-forward-full.json",
-            "durable_postgres",
-        ),
-        (
-            "examples/binance-multitimeframe/walk-forward-growth-optimal.json",
-            "durable_postgres",
-        ),
-    ),
-)
-def test_maintained_configs_pin_sealed_ledger_mode(
-    path: str,
-    expected: str,
-) -> None:
-    payload = json.loads(Path(path).read_text(encoding="utf-8"))
-    assert payload["sealed_test_ledger_mode"] == expected
-'''
-if "test_maintained_configs_pin_sealed_ledger_mode" in source:
-    raise SystemExit("maintained config contract already exists")
-test_path.write_text(source.rstrip() + addition + "\n", encoding="utf-8")
