+++ examples/binance-multitimeframe/walk-forward-full.json
@@ -221,5 +221,6 @@
     "step_bars": 2880,
     "test_bars": 2880,
     "train_bars": 20000
-  }
+  },
+  "sealed_test_ledger_mode": "durable_postgres"
 }
