+++ examples/binance-multitimeframe/walk-forward-growth-optimal.json
@@ -222,5 +222,6 @@
     "step_bars": 2880,
     "test_bars": 2880,
     "train_bars": 20000
-  }
+  },
+  "sealed_test_ledger_mode": "durable_postgres"
 }
